No longer require 128 bit atomics for gc_lock (MLDB-1909)
